+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add new `failureHandler` to show useful information about `HelmReleases` that are not ready.
+
 ## [1.40.0] - 2025-08-29
 
 ### Fixed